# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.RegisterInput:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  model.Supplier:
    properties:
      address:
        type: string
      contact_name:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  model.SupplierInput:
    properties:
      address:
        type: string
      contact_name:
        maxLength: 100
        type: string
      email:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - name
    type: object
...
  model.UpdateUserInput:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: CRM Backend
  description: |-
    Бэкенд CRM-системы для предприятия оптово-розничной торговли:
    управление продуктами, категориями, поставщиками, заявками на закупку,
    счетами-фактурами, аналитикой и отчётами.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: CRM API
  version: "1.0"
paths:
...
  /suppliers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Supplier'
            type: array
      security:
      - BearerAuth: []
      summary: Список поставщиков
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные поставщика
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SupplierInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать поставщика
      tags:
      - suppliers
  /suppliers/{id}:
    delete:
      parameters:
      - description: ID поставщика
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить поставщика
      tags:
      - suppliers
    get:
      parameters:
      - description: ID поставщика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Supplier'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить поставщика
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      parameters:
      - description: ID поставщика
        in: path
        name: id
        required: true
        type: integer
      - description: Данные поставщика
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SupplierInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Supplier'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновить поставщика
      tags:
      - suppliers
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя (можно задать роль)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Меняет пароль и/или роль
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновить пользователя
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...